virtq: Optimize LuaJIT behavior [DRAFT]
Simple bit: don't assign the avail ring to a variable ("avail")
because this can cause an allocation for the pointer.

Harder bit: change the way direct vs. indirect descriptors are
resolved. Direct descriptors are used unless/until
enable_indirect_descriptors() is called.

Problem: enable_indirect_descriptors() should be called by net_device
when appropriate but currently it is not.