core.memory: Use shm (POSIX shared memory) to allocate HugeTLB
Quoting the new comment in memory.c:

The HugeTLB page is allocated and mapped using the shm (shared
memory) API. This API makes it easy to remap the page to a new
virtual address after we resolve the physical address.

There are two other APIs for allocating HugeTLB pages but they do
not work as well:

  mmap() anonymous page with MAP_HUGETLB: cannot remap the address
  after allocation because Linux mremap() does not seem to work on
  HugeTLB pages.

  mmap() with file-backed MAP_HUGETLB: the file has to be on a
  hugetlbfs mounted filesystem and that is not necessarily
  available.

Happily the shm API is happy to remap a HugeTLB page with an
additional call to shmat() and does not depend on hugetlbfs.

Further reading:
  https://www.kernel.org/doc/Documentation/vm/hugetlbpage.txt
  http://stackoverflow.com/questions/27997934/mremap2-with-hugetlb-to-change-virtual-address